+++ app/src/main/java/de/halfbit/edgetoedge/sample/MainAdapter.kt
@@ -8,16 +8,18 @@ import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import de.halfbit.edgetoedge.sample.examples.SplashScreenFragment
-import de.halfbit.edgetoedge.sample.examples.ToolbarWithScrollableContent
-import de.halfbit.edgetoedge.sample.examples.ToolbarWithScrollableContentAndFab
+import de.halfbit.edgetoedge.sample.examples.ToolbarWithScrollableContentFragment
+import de.halfbit.edgetoedge.sample.examples.ToolbarWithScrollableContentAndFabFragment
+import de.halfbit.edgetoedge.sample.examples.ViewpagerFragment
 
 class MainAdapter(private val onItemClicked: OnItemClicked) :
     RecyclerView.Adapter<ItemViewHolder>() {
 
     private var items = listOf(
         Item(R.string.splash) { SplashScreenFragment() },
-        Item(R.string.toolbar) { ToolbarWithScrollableContent() },
-        Item(R.string.toolbar_fab) { ToolbarWithScrollableContentAndFab() }
+        Item(R.string.toolbar) { ToolbarWithScrollableContentFragment() },
+        Item(R.string.toolbar_fab) { ToolbarWithScrollableContentAndFabFragment() },
+        Item(R.string.toolbar_viewpager) { ViewpagerFragment() }
     )
 
     override fun getItemCount(): Int = items.size