+++ app/src/main/java/de/halfbit/edgetoedge/sample/examples/ToolbarWithScrollableContentFragment.kt
@@ -6,15 +6,19 @@ import de.halfbit.edgetoedge.Edge
 import de.halfbit.edgetoedge.edgeToEdge
 import de.halfbit.edgetoedge.sample.BaseFragment
 import de.halfbit.edgetoedge.sample.R
-import kotlinx.android.synthetic.main.fragment_splash_screen.*
+import de.halfbit.edgetoedge.sample.commons.ImagesAdapter
+import kotlinx.android.synthetic.main.fragment_toolbar_with_scrollable_content.*
 
-class SplashScreenFragment : BaseFragment() {
-    override val layoutId: Int get() = R.layout.fragment_splash_screen
+class ToolbarWithScrollableContentFragment : BaseFragment() {
+    override val layoutId: Int get() = R.layout.fragment_toolbar_with_scrollable_content
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         edgeToEdge {
-            container.fit { Edge.Top + Edge.Bottom }
+            appbar.fit { Edge.Top }
+            recycler.fit { Edge.Bottom }
         }
+
+        recycler.adapter = ImagesAdapter(requireContext())
     }
 }