+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 
-# 0.1 (26.12.2019)
+# 0.2 (26.12.2019)
+* Added property `fit { consumeInsets: Boolean }` for suppressing insets consumption.
+* Added `Toolbar + ViewPager in ConstraintLayout` sample screen.
 
+# 0.1 
 * Initial release